// src/ongeki/model/userMusic.ts
export interface UserMusicDetailItem {
  musicId: number;
  level: number;
  playCount: number;
  techScoreMax: number;
  techScoreRank: number;
  battleScoreMax: number;
  battleScoreRank: number;
  maxComboCount: number;
  maxOverKill: number;
  maxTeamOverKill: number;
  isFullBell: boolean;
  isFullCombo: boolean;
  isAllBreake: boolean;
  isLock: boolean;
  clearStatus: number;
  isStoryWatched: boolean;
}